+++ inst/NEWS.md
@@ -8,6 +8,7 @@
 ## BUG FIXES
 - Bug fix in `oncoplot` for drawing borders. [958](https://github.com/PoisonAlien/maftools/issues/958)
 - Bug fix in `plotSignatures` for hardcoded axis limits. [949](https://github.com/PoisonAlien/maftools/issues/949)
+- Bug fix in `mafSurvival` legend when samples argument is give. [937](https://github.com/PoisonAlien/maftools/issues/937)
 - Bug fix in `subsetMaf` while handling only CNV events. [908](https://github.com/PoisonAlien/maftools/issues/908)
 - Error handling when no deep/shallow CNV events found. [899](https://github.com/PoisonAlien/maftools/issues/899)
 - Bug fix in `oncoplot` for duplicated values in gene list. [889](https://github.com/PoisonAlien/maftools/issues/889)
@@ -15,7 +16,9 @@
 ## ENHANCEMENTS
 - Improved `annovarToMaf` with better handling of indels and `Variant_Type`. Issue: [940](https://github.com/PoisonAlien/maftools/issues/940)
 - Include absolute contribution of each signature in `extractSignatures` output. Issue: [939](https://github.com/PoisonAlien/maftools/issues/939)
+- Added `tsbToPIDs` for custom names in `oncoplot`. Issue: Issue: [936](https://github.com/PoisonAlien/maftools/issues/936)
 - Added `DSEL` protein to the database. Issue: [933](https://github.com/PoisonAlien/maftools/issues/933)
+- Added `MUC3A` protein to the database. Issue: [932](https://github.com/PoisonAlien/maftools/issues/932)
 - Added `showOnlyPathway` argument to `oncoplot`
 - Added `pathdb` argument to `PlotOncogenicPathways`. Issue: [923](https://github.com/PoisonAlien/maftools/issues/923)
 - Emit warnings when fishers test can not be performed during `somaticInteractions`. Issue: [921](https://github.com/PoisonAlien/maftools/issues/921)